* master: (912 commits)
  Update redirects.txt
  Added redirect for Spark guide link
  client: log when server list changes
  docs: mention regression in task config validation
  fix update to changelog
  update CHANGELOG with datacenter config validation https://github.com/hashicorp/nomad/pull/5665
  typo: "atleast" -> "at least"
  implement nomad exec for rkt
  docs: fixed typo
  use pty/tty terminology similar to github.com/kr/pty
  vendor github.com/kr/pty
  drivers: implement streaming exec for executor based drivers
  executors: implement streaming exec
  executor: scaffolding for executor grpc handling
  client: expose allocated memory per task
  client improve a comment in updateNetworks
  stalebot: Add 'thinking' as an exempt label (#5684)
  Added Sparrow link
  update links to use new canonical location
  Add redirects for restructing done in GH-5667
  ...

+++ scheduler/spread.go
@@ -28,7 +28,7 @@ type SpreadIterator struct {
 
 	// sumSpreadWeights tracks the total weight across all spread
 	// stanzas
-	sumSpreadWeights int
+	sumSpreadWeights int32
 
 	// hasSpread is used to early return when the job/task group
 	// does not have spread configured
@@ -43,7 +43,7 @@ type SpreadIterator struct {
 type spreadAttributeMap map[string]*spreadInfo
 
 type spreadInfo struct {
-	weight        int
+	weight        int8
 	desiredCounts map[string]float64
 }
 
@@ -246,7 +246,7 @@ func (iter *SpreadIterator) computeSpreadInfo(tg *structs.TaskGroup) {
 			si.desiredCounts[implicitTarget] = remainingCount
 		}
 		spreadInfos[spread.Attribute] = si
-		iter.sumSpreadWeights += spread.Weight
+		iter.sumSpreadWeights += int32(spread.Weight)
 	}
 	iter.tgSpreadInfo[tg.Name] = spreadInfos
 }
